rename equation folder
