+++ lib/solveEquation/Equation.js
