+++ lib/solveEquation/EquationOperations.js
