+++ lib/solveEquation/EquationStatus.js
