+++ lib/solveEquation/solveEquation.js
