+++ test/solveEquation.test.js
@@ -4,7 +4,7 @@ const assert = require('assert');
 const math = require('mathjs');
 
 const MathChangeTypes = require('../lib/MathChangeTypes');
-const solveEquation = require('../lib/equation/solveEquation');
+const solveEquation = require('../lib/solveEquation/solveEquation');
 
 const NO_STEPS = 'no-steps';
